Skip Selenium tests if not available
If either the Python selenium module or the necessary webdriver
is not available, skip the Selenium tests.  This allows running
the test suite in integration test environments where Selenium
is not available.
