+++ tests/test_fe_users.py
@@ -1,6 +1,6 @@
 from fixtures import graph, groups, permissions, session, standard_graph, users  # noqa: F401
-from fixtures import async_server, browser  # noqa: F401
 from fixtures import fe_app as app  # noqa: F401
+from fixtures_selenium import async_server, browser  # noqa: F401
 from pages import UserViewPage
 from plugins import group_ownership_policy
 from url_util import url
